+++ src/main/java/se/bjurr/sscc/ResultsCallable.java
@@ -1,0 +1,10 @@
+package se.bjurr.sscc;
+
+import com.google.common.annotations.VisibleForTesting;
+
+public class ResultsCallable {
+ @VisibleForTesting
+ public void report(boolean isAccepted, String summary, String message) {
+
+ }
+}