+++ src/main/java/se/bjurr/sscc/SsccUserAdminService.java
@@ -1,0 +1,9 @@
+package se.bjurr.sscc;
+
+import java.util.Map;
+
+import com.atlassian.stash.user.DetailedUser;
+
+public interface SsccUserAdminService {
+ Map<String, DetailedUser> getStashUsers();
+}